chief complaint word crap + more decimal changes

+++ app/femr/ui/models/search/CreateEncounterViewModel.java
@@ -3,14 +3,14 @@ package femr.ui.models.search;
 import java.util.List;
 
 public class CreateEncounterViewModel {
-    private Float bloodPressureSystolic;
-    private Float bloodPressureDiastolic;
-    private Float heartRate;
+    private Integer bloodPressureSystolic;
+    private Integer bloodPressureDiastolic;
+    private Integer heartRate;
     private Float temperature;
-    private Float respiratoryRate;
+    private Integer respiratoryRate;
     private Float oxygenSaturation;
-    private Float heightFeet;
-    private Float heightInches;
+    private Integer heightFeet;
+    private Integer heightInches;
     private Float weight;
     private String firstName;
     private String lastName;
@@ -25,27 +25,27 @@ public class CreateEncounterViewModel {
     private List<String> perscribList;
 
 
-    public Float getBloodPressureSystolic() {
+    public Integer getBloodPressureSystolic() {
         return bloodPressureSystolic;
     }
 
-    public void setBloodPressureSystolic(Float bloodPressureSystolic) {
+    public void setBloodPressureSystolic(Integer bloodPressureSystolic) {
         this.bloodPressureSystolic = bloodPressureSystolic;
     }
 
-    public Float getBloodPressureDiastolic() {
+    public Integer getBloodPressureDiastolic() {
         return bloodPressureDiastolic;
     }
 
-    public void setBloodPressureDiastolic(Float bloodPressureDiastolic) {
+    public void setBloodPressureDiastolic(Integer bloodPressureDiastolic) {
         this.bloodPressureDiastolic = bloodPressureDiastolic;
     }
 
-    public Float getHeartRate() {
+    public Integer getHeartRate() {
         return heartRate;
     }
 
-    public void setHeartRate(Float heartRate) {
+    public void setHeartRate(Integer heartRate) {
         this.heartRate = heartRate;
     }
 
@@ -57,11 +57,11 @@ public class CreateEncounterViewModel {
         this.temperature = temperature;
     }
 
-    public Float getRespiratoryRate() {
+    public Integer getRespiratoryRate() {
         return respiratoryRate;
     }
 
-    public void setRespiratoryRate(Float respiratoryRate) {
+    public void setRespiratoryRate(Integer respiratoryRate) {
         this.respiratoryRate = respiratoryRate;
     }
 
@@ -73,19 +73,19 @@ public class CreateEncounterViewModel {
         this.oxygenSaturation = oxygenSaturation;
     }
 
-    public Float getHeightFeet() {
+    public Integer getHeightFeet() {
         return heightFeet;
     }
 
-    public void setHeightFeet(Float heightFeet) {
+    public void setHeightFeet(Integer heightFeet) {
         this.heightFeet = heightFeet;
     }
 
-    public Float getHeightInches() {
+    public Integer getHeightInches() {
         return heightInches;
     }
 
-    public void setHeightInches(Float heightInches) {
+    public void setHeightInches(Integer heightInches) {
         this.heightInches = heightInches;
     }
